News for version 0.9.4 [released: 2014-01-02] *  BUG ID  6: autogen.sh fails on Mac OS X 10.6.8 *  BUG ID  7: make fails on missing file in .deps directory (Mac OS X 10.6.8) *  BUG ID 10: Install step does not install supersonic/cursor/base/cursor_transformer.h *  BUG ID 11: shared_ptr declared in global scope *  Missing field initialization in RowHashSet. Bug reported by: st xiao https://groups.google.com/forum/?fromgroups=#!topic/supersonic-query-engine/UPqZLCuHxE0
Features:
* Significant cleanup of dependencies:
GOOD news: Supersonic builds on OSX and should be much easier to build it on other
systems.
BAD news:  C++0x compatible compiler is required (e.g. >=gcc-4.6).
* Added support for enums to Supersonic:
* adds core support for the enum type, that consists of (id, value) pairs.
* adds support for parsing enums (by name) from strings.
* adds support for printing enums to strings (when dumping views for
debugging and testing).
* Extended aggegation support: Adding the functionality for limiting number of
rows in the result by giving a bound on the unique number of key rows seen in
input order. All the key combinations which doesn't fall in this bound get
aggregated in one single row at last hash index.
Affected APIs:
supersonic/cursor/core/aggregate.h
supersonic/cursor/infrastructure/row_hash_set.h
* Fixed potential undefined-behaviour issue in:
supersonic/expression/core/elementary_bound_expressions.cc
* Fixed bug in supersonic/expression/core/elementary_bound_expressions.cc
* hash-join optimization: for empty left-hand-side, don't scan the
right-hand-side.
* Fixed Permutation::permutation() behavior in case of empty permutation.

+++ supersonic/base/infrastructure/tuple_schema.cc
@@ -26,7 +26,7 @@ EnumDefinition::Rep::Rep(BufferAllocator* buffer_allocator)
       arena_(buffer_allocator, kInitialArenaBufferSize,
              kMaxArenaBufferSize) {}
 
-FailureOrVoid EnumDefinition::Rep::Add(const int64 number, StringPiece name) {
+FailureOrVoid EnumDefinition::Rep::Add(const int32 number, StringPiece name) {
   if (name_to_number_.count(name) > 0) {
     THROW(new Exception(
         ERROR_DUPLICATE_ENUM_VALUE_NAME,
@@ -60,7 +60,7 @@ void EnumDefinition::Rep::CopyFrom(const EnumDefinition::Rep& other) {
   }
 }
 
-FailureOr<StringPiece> EnumDefinition::Rep::NumberToName(int64 number) const {
+FailureOr<StringPiece> EnumDefinition::Rep::NumberToName(int32 number) const {
   auto iterator = number_to_name_.find(number);
   if (iterator == number_to_name_.end()) {
     THROW(new Exception(
@@ -70,7 +70,7 @@ FailureOr<StringPiece> EnumDefinition::Rep::NumberToName(int64 number) const {
   return Success(iterator->second);
 }
 
-FailureOr<int64> EnumDefinition::Rep::NameToNumber(StringPiece name) const {
+FailureOr<int32> EnumDefinition::Rep::NameToNumber(StringPiece name) const {
   auto iterator = name_to_number_.find(name);
   if (iterator == name_to_number_. end()) {
     THROW(new Exception(
@@ -105,7 +105,7 @@ FailureOrVoid EnumDefinition::Rep::VerifyEquals(const EnumDefinition::Rep& a,
 
 EnumDefinition::EnumDefinition() {}
 
-FailureOrVoid EnumDefinition::AddEntry(const int64 number, StringPiece name) {
+FailureOrVoid EnumDefinition::AddEntry(const int32 number, StringPiece name) {
   // Lazy, to avoid creating for non-enums.
   if (rep_.get() == NULL) {
     // NOTE(user): uses the unbounded (heap) allocator for now. Given that
@@ -124,7 +124,7 @@ size_t EnumDefinition::entry_count() const {
   return rep_.get() == NULL ? 0 : rep_->entry_count();
 }
 
-FailureOr<StringPiece> EnumDefinition::NumberToName(int64 number) const {
+FailureOr<StringPiece> EnumDefinition::NumberToName(int32 number) const {
   if (rep_.get() == NULL) {
     THROW(new Exception(ERROR_UNDEFINED_ENUM_VALUE_NUMBER,
                         "The enum is empty"));
@@ -133,7 +133,7 @@ FailureOr<StringPiece> EnumDefinition::NumberToName(int64 number) const {
   }
 }
 
-FailureOr<int64> EnumDefinition::NameToNumber(StringPiece name) const {
+FailureOr<int32> EnumDefinition::NameToNumber(StringPiece name) const {
   if (rep_.get() == NULL) {
     THROW(new Exception(ERROR_UNDEFINED_ENUM_VALUE_NAME, "The enum is empty"));
   } else {
